+++ src/Phan/Analysis/PostOrderAnalysisVisitor.php
@@ -575,9 +575,14 @@ class PostOrderAnalysisVisitor extends AnalysisVisitor
                 $this->context,
                 $node->children['default']
             );
-
-            $variable->setUnionType($default_type);
+        } else {
+            $default_type = NullType::instance(false)->asUnionType();
         }
+
+        $variable->setUnionType($default_type);
+        // TODO: Probably not true in a loop?
+        // TODO: Expand this to assigning to variables? (would need to make references invalidate that, and skip this in the global scope)
+        $variable->enablePhanFlagBits(\Phan\Language\Element\Flags::IS_CONSTANT_DEFINITION);
 
         // Note that we're not creating a new scope, just
         // adding variables to the existing scope
@@ -1807,6 +1812,16 @@ class PostOrderAnalysisVisitor extends AnalysisVisitor
                 continue;
             }
             // Don't bother recursing more than one level to iterate over possible types.
+            if ($elem->kind === \ast\AST_UNPACK) {
+                // Could optionally recurse to better analyze `yield [...SOME_EXPRESSION_WITH_MIX_OF_VALUES]`
+                yield $elem->lineno => UnionTypeVisitor::unionTypeFromNode(
+                    $this->code_base,
+                    $context,
+                    $elem,
+                    true
+                );
+                continue;
+            }
             $value_node = $elem->children['value'];
             if ($value_node instanceof Node) {
                 yield $elem->lineno => UnionTypeVisitor::unionTypeFromNode(