+++ tests/Phan/CLITest.php
@@ -95,6 +95,9 @@ final class CLITest extends BaseTest
         }
         \ksort($changed);
         \ksort($expected_changed_options);
+        if (!\array_key_exists('color_issue_messages', $expected_changed_options)) {
+            unset($changed['color_issue_messages']);
+        }
         $this->assertSame($expected_changed_options, $changed);
 
         $this->assertSame(['src' . \DIRECTORY_SEPARATOR . 'empty.php'], $cli->getFileList());
@@ -167,6 +170,10 @@ final class CLITest extends BaseTest
                     'language-server-enable-hover' => false,
                     'language-server-enable-completion' => false,
                 ],
+            ],
+            [
+                ['language_server_min_diagnostics_delay_ms' => 100.0],
+                ['language-server-min-diagnostics-delay-ms' => '100'],
             ],
             [
                 [